+++ src/validator/Validator.js
@@ -585,7 +585,7 @@ class ValidationChecker {
 
           inputMask = inputMask ? getInputMask(inputMask) : null;
 
-          if (value && inputMask) {
+          if (value && inputMask && !component.skipMaskValidation) {
             return matchInputMask(value, inputMask);
           }
 
@@ -802,7 +802,7 @@ class ValidationChecker {
       return result ? {
         message: _.get(result, 'message', result),
         level: _.get(result, 'level') === 'warning' ? 'warning' : 'error',
-        path: getArrayFromComponentPath(component.calculatedPath || ''),
+        path: getArrayFromComponentPath(component.path || ''),
         context: {
           validator: validatorName,
           setting,