+++ test/formtest/index.js
@@ -11,6 +11,8 @@ const uniqueApiKeys = require('./uniqueApiKeys.json');
 const uniqueApiKeysLayout = require('./uniqueApiKeysLayout.json');
 const uniqueApiKeysSameLevel = require('./uniqueApiKeysSameLevel.json');
 const validationOnBlur = require('./validationOnBlur.json');
+const calculateValueWithManualOverride = require('./calculateValueWithManualOverride.json');
+const displayAsModalEditGrid = require('./displayAsModalEditGrid.json');
 
 module.exports = {
   advanced,
@@ -25,5 +27,7 @@ module.exports = {
   uniqueApiKeys,
   uniqueApiKeysLayout,
   uniqueApiKeysSameLevel,
-  validationOnBlur
+  validationOnBlur,
+  calculateValueWithManualOverride,
+  displayAsModalEditGrid
 };